+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Generic mappers (`Map1`, ...) have methods `NewEntities`, `NewEntitiesWith` and `RemoveEntities` for batch operations (#151)
 * Batch-creation methods (ID-based and generic) have variants like `NewEntitiesQuery` that return a query over the created entities (#152)
 * Notification during batch-creation is delayed until the resp. query is closed (#157)
+* Match-remove methods (`RemoveEntities()`) return the number of removed entities (#173)
 
 ### Performance
 
@@ -33,6 +34,10 @@
 ### Bugfixes
 
 * Archetype storage buffers are "zeroed" when removing entities, to allow GC on pointers and slices in components (#147)
+
+### Documentation
+
+* Adds an example for batch-creation and batch-removal of entities (#173)
 
 ### Other